+++ migration/changelog.md
@@ -4,6 +4,14 @@
 
 ### User 欄位變動
 
-* issue link
+
 
 ### Recipe 欄位變動
+
+
+## v2
+
+### Feed 欄位變動
+
+* 新增 createdTime 紀錄 feed 在 facebook 的建立時間
+* 新增 publish 紀錄 feed 是否發布